RavenDB-22109: Ensure `VerifyLicense` is invoked in all `RavenServer` instantiations
