+++ src/Raven.Server/RavenServer.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Buffers;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -33,6 +32,7 @@ using Raven.Client.Exceptions.Security;
 using Raven.Client.Extensions;
 using Raven.Client.Http;
 using Raven.Client.Json.Serialization;
+using Raven.Client.Properties;
 using Raven.Client.ServerWide.Operations.Certificates;
 using Raven.Client.ServerWide.Tcp;
 using Raven.Client.Util;
@@ -151,6 +151,9 @@ namespace Raven.Server
             _tcpLogger = LoggingSource.Instance.GetLogger<RavenServer>("Server/TCP");
             _externalCertificateValidator = new ExternalCertificateValidator(this, Logger);
             _tcpContextPool = new JsonContextPool(Configuration.Memory.MaxContextSizeToKeep);
+
+            // doing this before the schema upgrade to allow to downgrade in case we cannot start the server
+            BeforeSchemaUpgrade = x => VerifyLicense(x, ServerStore);
         }
 
         public TcpListenerStatus GetTcpServerStatus()
@@ -2964,5 +2967,80 @@ namespace Raven.Server
             ArrayPool<byte>.Shared.Return(buffer);
         }
 
+        internal static void VerifyLicense(StorageEnvironment storageEnvironment, ServerStore serverStore)
+        {
+            using (var contextPool = new TransactionContextPool(storageEnvironment, serverStore.Configuration.Memory.MaxContextSizeToKeep))
+            {
+                var license = serverStore.LoadLicense(contextPool);
+                if (license == null)
+                    return;
+
+                var licenseStatus = LicenseManager.GetLicenseStatus(license);
+                if (licenseStatus.Expiration >= RavenVersionAttribute.Instance.ReleaseDate)
+                    return;
+
+                string licenseJson = null;
+                var fromPath = false;
+                if (string.IsNullOrEmpty(serverStore.Configuration.Licensing.License) == false)
+                {
+                    licenseJson = serverStore.Configuration.Licensing.License;
+                }
+                else if (File.Exists(serverStore.Configuration.Licensing.LicensePath.FullPath))
+                {
+                    try
+                    {
+                        licenseJson = File.ReadAllText(serverStore.Configuration.Licensing.LicensePath.FullPath);
+                        fromPath = true;
+                    }
+                    catch
+                    {
+                        // expected
+                    }
+                }
+
+                var errorMessage = $"Cannot start the RavenDB server because the expiration date of current license ({FormattedDateTime(licenseStatus.Expiration ?? DateTime.MinValue)}) " +
+                                   $"is before the release date of this version ({FormattedDateTime(RavenVersionAttribute.Instance.ReleaseDate)})";
+
+                string expiredLicenseMessage = "";
+                if (string.IsNullOrEmpty(licenseJson) == false)
+                {
+                    if (LicenseHelper.TryDeserializeLicense(licenseJson, out License localLicense))
+                    {
+                        var localLicenseStatus = LicenseManager.GetLicenseStatus(localLicense);
+                        if (localLicenseStatus.Expiration >= RavenVersionAttribute.Instance.ReleaseDate)
+                        {
+                            serverStore.LicenseManager.OnBeforeInitialize += () => serverStore.LicenseManager.TryActivateLicenseAsync(throwOnActivationFailure: serverStore.Server.ThrowOnLicenseActivationFailure).Wait(serverStore.ServerShutdown);
+                            return;
+                        }
+
+                        var configurationKey =
+                            fromPath ? RavenConfiguration.GetKey(x => x.Licensing.LicensePath) : RavenConfiguration.GetKey(x => x.Licensing.License);
+                        expiredLicenseMessage = localLicense.Id == license.Id
+                            ? ". You can update current license using the setting.json file"
+                            : $". The license '{localLicense.Id}' obtained from '{configurationKey}' with expiration date of '{FormattedDateTime(localLicenseStatus.Expiration ?? DateTime.MinValue)}' is also expired.";
+                    }
+                    else
+                    {
+                        errorMessage += ". Could not parse the license from setting.json file.";
+                        throw new LicenseExpiredException(errorMessage);
+                    }
+                }
+
+                var licenseStorage = new LicenseStorage();
+                licenseStorage.Initialize(storageEnvironment, contextPool);
+
+                var buildInfo = licenseStorage.GetBuildInfo();
+                if (buildInfo != null)
+                    errorMessage += $" You can downgrade to the latest build that was working ({buildInfo.FullVersion})";
+                if (string.IsNullOrEmpty(expiredLicenseMessage) == false)
+                    errorMessage += expiredLicenseMessage;
+                throw new LicenseExpiredException(errorMessage);
+
+                static string FormattedDateTime(DateTime dateTime)
+                {
+                    return dateTime.ToString("dd MMMM yyyy");
+                }
+            }
+        }
     }
 }
